requirements.txt
alembic==1.8.1
apispec==5.2.2
attrs==22.1.0
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
ecdsa==0.18.0
filelock==3.8.0
flasgger==0.9.5
Flask==2.2.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
identify==2.5.5
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==2.0.0
marshmallow-sqlalchemy==0.28.1
mistune==2.0.4
nodeenv==1.7.0
packaging==21.3
platformdirs==2.5.2
pre-commit==2.20.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.21.0
python-jose==3.3.0
PyYAML==6.0
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.41
toml==0.10.2
virtualenv==20.16.5
Werkzeug==2.2.2
